requirements.txt
backoff==1.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
docopt==0.6.2
googleapis-common-protos==1.56.4
grpcio==1.47.0
idna==3.3
opentelemetry-api==1.12.0rc2
opentelemetry-distro==0.32b0
opentelemetry-exporter-otlp==1.12.0rc2
opentelemetry-exporter-otlp-proto-grpc==1.12.0rc2
opentelemetry-exporter-otlp-proto-http==1.12.0rc2
opentelemetry-instrumentation==0.32b0
opentelemetry-proto==1.12.0rc2
opentelemetry-sdk==1.12.0rc2
opentelemetry-semantic-conventions==0.32b0
pipreqs==0.4.11
protobuf==3.20.1
requests==2.28.1
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.10
wrapt==1.14.1
yarg==0.1.9